refactor(analysis): extract conversation read API into its own module (#416)
Final slice of the analysis decomposition (read-side only, by design).
ConversationEntity/PacketEntity are a shared kernel read by ~9 modules,
so they stay in the analysis core; only the conversation browse/query API
is extracted.

- New ConversationQueryService (com.tracepcap.conversation.service) holds
  the conversation list/facets/detail/CSV+PCAP-export logic carved out of
  AnalysisService (~450 lines); AnalysisService drops from 1174 to ~710
- ConversationsController -> com.tracepcap.conversation.controller, now
  calling ConversationQueryService (and SessionReconstructionService for
  /session); path /conversations unchanged
- Conversation DTOs, ConversationEntity/PacketEntity, repositories, and
  getSecurityAlerts (returns ConversationResponse) stay in analysis core
- Removed now-unused imports and the orphaned ipGeoInfoRepository field

Dependency direction is conversation -> analysis (consumes the core
entity/repo/DTOs), never the reverse. OpenAPI contract unchanged.
Verified with mvn verify (JUnit + Testcontainers).

Closes #416.
